Crate mithril_metric

metrics module. This module contains the tools to create a metrics service and a metrics server.

Re-exports

pub use metric::*;

Modules

helper

Helper to create a metric service.

metric

This module contains wrapper to prometheus metrics for use in a metrics service.

Macros

build_metrics_service

Create a MetricService.

Structs

MetricsServer

The MetricsServer is responsible for exposing the metrics of the signer.

Traits

MetricsServiceExporter

Metrics service exporter gives the possibility of exporting metrics.